Implementado: Buscar os simbolos da categoria selecionada e exibir em um novo Dialog.
Falta:
Corrigir layout do Dialog de simbolos
Selecionar o simbolo e vincular ele ao imageView
Encontrar uma maneira de salvar esses simbolos e gravar na base
Sincronizar os planos do Euroku.

+++ Tagarela/src/br/com/furb/tagarela/interfaces/CategoryTypeListener.java
@@ -1,5 +1,5 @@
 package br.com.furb.tagarela.interfaces;
 
 public interface CategoryTypeListener {
-	public void onCategoryReturnValue(Long categoryID);
+	public void onCategoryReturnValue(int categoryID);
 }
